+++ tests/testapp/settings.py
@@ -57,7 +57,7 @@ MIDDLEWARE = [
     "django.middleware.clickjacking.XFrameOptionsMiddleware",
 ]
 
-USE_CUSTOM_SITE = bool(os.environ.get('USE_CUSTOM_SITE', False))
+USE_CUSTOM_SITE = bool(os.environ.get("USE_CUSTOM_SITE", False))
 if USE_CUSTOM_SITE:
     FEINCMS3_SITES_SITE_MODEL = "testapp.CustomSite"
 else: