+++ tests/testapp/test_feincms3.py
@@ -11,7 +11,7 @@ from feincms3_sites.middleware import set_current_site
 from feincms3_sites.models import Site
 from feincms3_sites.utils import get_site_model
 
-from .models import Article, Page, CustomSite
+from .models import Article, CustomSite, Page
 
 
 def zero_management_form_data(prefix):
@@ -514,7 +514,8 @@ class SiteModelDeclaredTest(TestCase):
 
         if settings.USE_CUSTOM_SITE:
             self.test_site = CustomSite.objects.create(
-                host="testserver", is_default=True, title='Test Site')
+                host="testserver", is_default=True, title="Test Site"
+            )
         else:
             self.test_site = Site.objects.create(host="testserver", is_default=True)